# cpu: matmul: do not use redundant kernel