It was found during implementing [MFDNN-13994](https://jira.devtools.intel.com/browse/MFDNN-13994) that GeMM-Matmul works very slowly in some cases for fp32 tensors. The reason was in creation of a kernel that processes post-ops in cases when this could be avoided. Then the kernel run using all available threads which introduced significant overhead.
The case of bf16 tensors is not reproducible using RNN but is similar.

This PR fixes this performance issue.

Measured on SPR CPU.

fp32
```
main branch:
$ OMP_NUM_THREADS=48 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=f32 --attr-post-ops=sum:2.0 256x1:1x256
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:f32,,--mode=P --matmul --attr-post-ops=sum:2 256x1:1x256,0.000131072,0.805176,0.0510254,2.56876,0.0607429,2.15782
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:f32 : 1 (100%)                                  |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.0510254 avg(ms):0.0607429

PR:
$ OMP_NUM_THREADS=48 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=f32 --attr-post-ops=sum:2.0 256x1:1x256
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:f32,,--mode=P --matmul --attr-post-ops=sum:2 256x1:1x256,0.000131072,0.705811,0.00561523,23.3422,0.0063097,20.7731
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:f32 : 1 (100%)                                  |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.00561523 avg(ms):0.0063097
```

bf16
```
main branch:
$ OMP_NUM_THREADS=48 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2.0 1x16:16x16
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:bf16,,--mode=P --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2 1x16:16x16,5.12e-07,0.951904,0.0124512,0.0411206,0.016142,0.0317185
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:bf16 : 1 (100%)                                 |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.0124512 avg(ms):0.016142

PR:
OMP_NUM_THREADS=48 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2.0 1x16:16x16
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:bf16,,--mode=P --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2 1x16:16x16,5.12e-07,0.7229,0.000244141,2.09715,0.000514927,0.994315
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:bf16 : 1 (100%)                                 |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.000244141 avg(ms):0.000514927
```


### Update

Multithreading overhead is significant for small shapes. For larger shapes, performance is also improved.

Here are some test results for larger shapes run on ONE thread.

fp32
```
main branch:
$ OMP_NUM_THREADS=1 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=f32 --attr-post-ops=sum:2.0 4096x8:8x4096
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:f32,,--mode=P --matmul --attr-post-ops=sum:2 4096x8:8x4096,0.268435,0.888672,8.76514,30.6254,8.93805,30.0329
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:f32 : 1 (100%)                                  |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):8.76514 avg(ms):8.93805

PR:
$ OMP_NUM_THREADS=1 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=f32 --attr-post-ops=sum:2.0 4096x8:8x4096
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:f32,,--mode=P --matmul --attr-post-ops=sum:2 4096x8:8x4096,0.268435,0.752441,4.94702,54.262,5.07083,52.9372
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:f32 : 1 (100%)                                  |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):4.94702 avg(ms):5.07083
```


bf16
```
main branch:
$ OMP_NUM_THREADS=1 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2.0 1x1024:1024x1024
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:bf16,,--mode=P --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2 1x1024:1024x1024,0.00209715,0.625,0.0424805,49.3674,0.0438958,47.7757
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:bf16 : 1 (100%)                                 |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.0424805 avg(ms):0.0438958

PR:
$ OMP_NUM_THREADS=1 ./tests/benchdnn/benchdnn --mode=p --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2.0 1x1024:1024x1024
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit:bf16,,--mode=P --matmul --dt=bf16:bf16:f32 --stag=ab --wtag=ab --dtag=ab --runtime_dims_masks=2:1 --attr-post-ops=sum:2 1x1024:1024x1024,0.00209715,0.525146,0.0354004,59.2409,0.0366045,57.2922
============================================================
= Implementation statistics (--summary=no-impl to disable) =
============================================================
| gemm:jit:bf16 : 1 (100%)                                 |
============================================================
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):0.0354004 avg(ms):0.0366045

```