@asimonov1 2025-09-22
make test
disable test_device_gpu
disable build_gpu_runtime_ocl
disable build_gpu_runtime_sycl
disable benchdnn_all
enable benchdnn_matmul
set test_scope=NIGHTLY