@asimonov1 2025-09-26
> @asimonov1 out of curiosity since it is unclear to me just looking at the code change. Is the performance gain due to handling the post-op as part of the GEMM or is this somehow reducing the thread overhead?

Both. I found it when tested RNN, and the main overhead was introduced by multithreading (RNN uses relatively small shapes). I have done some more tests, and the performance is also better without multithreading. I updated the first comment with some results.